added youtube api
after adding youtube link to input field, you go to the next page that displays all the data from the video, so if you want to edit it you can.

TODO:
after you're done editing you press on 'upload' and it stores the upload in the database

Current error: The GET method is not supported for this route. Supported methods: POST.